Extract chunked transaction rollback out of chunked committer ()

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/tracer/DefaultTracer.java
@@ -255,7 +255,7 @@ public class DefaultTracer implements DatabaseTracer {
                 }
 
                 @Override
-                public void batchedRolledBack(int rolledBackBatches, long transactionId) {
+                public void batchedRolledBack(long rolledBackBatches, long transactionId) {
                     batchesRolledBack.add(rolledBackBatches);
                 }
             };
